lead-scorer/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
aiohttp==3.9.0
tqdm==4.66.1
jinja2==3.1.2
python-multipart==0.0.6
mangum==0.17.0